Avoid rendering JavaScript URLs as clickable links
To avoid running JavaScript taken from JSON files downloaded from the
internet in the client we now try to render URLs only as clickable links
if the protocol of the URL is safe, which more concretely means HTTP or
HTTPS. Other URLs are rendered as plain text. This commit only covers
links that are obviously treated as links to external resources.

To avoid duplication, this introduces a new component, SafeLink, that
takes the URL and optionally id and target attributes and renders a
suitable a-element if the protocol is safe and text otherwise.